#7: Moved certain methods from Fragment to ViewModel
